requirements.txt
asgiref==3.3.1
boto3==1.17.22
botocore==1.20.22
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.1
django-heroku==0.3.1
django-storages==1.11.1
gunicorn==20.0.4
jmespath==0.10.0
Pillow==8.1.2
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0